-fine-tuned hyperparameters -fixed max pooling -updated padding of convolutions -changed optimiser to Adam (from SGD) -updated README.md and log_notes.md -now using LeakyReLu, which provided slightly better results than Relu -fixed tensor concatenation being sent to 'device' -code now fully accounts for CUDA. You can copy and paste the script into a GPU and the script will automatically turn off non-GPU supported features. Note that Google Cloud servers do offer TensorBoard features; however I am using Paperspace Gradient, which does not have such support.

+++ run_CNN.py
@@ -9,38 +9,44 @@ import re
 import numpy as np
 import datetime
 import ErrorMetrics as metrics
-import ShowAerofoil as show
 import AerofoilDataset as AD
 import TitleSequence as Title
 import warnings
-from torch.utils.tensorboard import SummaryWriter
 import sys
 import subprocess
-from torch_lr_finder import LRFinder
+if not torch.cuda.is_available():
+    import ShowAerofoil as show
+    from torch_lr_finder import LRFinder
+    from torch.utils.tensorboard import SummaryWriter
 
+# device configuration
+torch.manual_seed(0)
+device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
 # output switches
-find_LR = False
+find_LR = False  # I don't think LR is giving accurate results. A suggested LR of 1 doesn't at all work for this model
 print_activations = False
-print_epoch = 25  # print output after n epochs (after doing all batches within epoch)
+print_epoch = 1  # print output after n epochs (after doing all batches within epoch)
 
 # hyper parameters
 hidden_layers = [300]
-convolutions = [6]  # , 16, 40, 120]
-num_epochs = 100
-bs = 25
-learning_rate = 2
+convolutions = [6, 16]
+num_epochs = 35
+bs = 5
+learning_rate = 0.01
 
 # file configuration
 time_of_run = datetime.datetime.now().strftime("D%d_%m_%Y_T%H_%M_%S")
 path = Path(__file__).parent
-train_dir = path / 'data' / 'out' / 'train'
-valid_dir = path / 'data' / 'out' / 'valid'
-test_dir = path / 'data' / 'out' / 'test'
-print_dir = path / 'print' / time_of_run
-writer = SummaryWriter(print_dir / 'TensorBoard_events')
-TB_process = subprocess.Popen(["tensorboard", f"--logdir={path / 'print'}"], stdout=open(os.devnull, 'w'),
-                              stderr=subprocess.STDOUT)  # {print_dir} to show just this run
+train_dir = path / ('storage/aerofoils' if torch.cuda.is_available() else 'data') / 'out' / 'train'
+valid_dir = path / ('storage/aerofoils' if torch.cuda.is_available() else 'data') / 'out' / 'valid'
+test_dir = path / ('storage/aerofoils' if torch.cuda.is_available() else 'data') / 'out' / 'test'
+print_dir = path / ('storage/aerofoils' if torch.cuda.is_available() else '') / 'print' / time_of_run
+print_dir.mkdir()
+if not torch.cuda.is_available():
+    writer = SummaryWriter(print_dir / 'TensorBoard_events')
+    TB_process = subprocess.Popen(["tensorboard", f"--logdir={path / 'print'}"], stdout=open(os.devnull, 'w'),
+                                  stderr=subprocess.STDOUT)  # {print_dir} to show just this run
 
 # write log file of architecture and for note taking
 with open(print_dir / "log.txt", 'w') as f:
@@ -66,10 +72,6 @@ num_channels = 1  # one channel for y coordinate (xy coordinates requires two ch
 Title.print_title([" ", "Convolutional neural network", "Outputs: Max ClCd @ angle",
                    "TensorBoardX running", f"Output directory: {'print/' + time_of_run}"])
 
-# device configuration
-torch.manual_seed(0)
-device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-
 # import datasets
 train_dataset = AD.AerofoilDataset(train_dir, num_channels, input_size, output_size,
                                    transform=transforms.Compose([AD.ToTensor()]))
@@ -84,51 +86,40 @@ valid_loader = DataLoader(dataset=valid_dataset, batch_size=bs, shuffle=True, nu
 test_loader = DataLoader(dataset=test_dataset, batch_size=1, shuffle=False, num_workers=4)  # bs = 1 for top_losses()
 
 # show aerofoils
-# show.show_aerofoil(writer, tensorboard=True, **train_dataset[0])
-# for i, batch in enumerate(valid_loader):
-#     show.show_aerofoil_batch(i, **batch)
+# if not torch.cuda.is_available():
+    # show.show_aerofoil(writer, tensorboard=True, **train_dataset[0])
+    # for i, batch in enumerate(valid_loader):
+    #     show.show_aerofoil_batch(i, **batch)
 
 
 class ConvNet(nn.Module):
     def __init__(self):
         super(ConvNet, self).__init__()
-        kernel_size = 2
-        stride = 2
-        dilation = 1  # default of nn.Conv1d = 1
-        padding = ((input_size - 1) * stride - input_size + kernel_size + (kernel_size - 1) * (dilation - 1)
-                   ) // 2
-        self.image_size = (input_size + 2 * padding - kernel_size - (kernel_size - 1) * (dilation - 1))\
-                          // stride + 1  # image size according to padding (as padding might not actually be an int)
+        kernel_size = 10
+        stride = 1  # no reduction in image size
+        # dilation = 1  # default of nn.Conv1d = 1
+
+        self.image_size = input_size
+        padding = (kernel_size - 1) // stride  # true if stride=1. Produces output size = input size after filter
 
         self.extractor = nn.Sequential(
-            nn.Conv1d(num_channels, convolutions[0], kernel_size, padding=padding, padding_mode='reflect'),  # 2 input channels (of 1D): x and y coords
-            nn.MaxPool1d(2, stride),
-            # nn.Conv1d(convolutions[0], convolutions[1], kernel_size, padding=padding, padding_mode='reflect'),
-            # nn.MaxPool1d(2, stride),
-            # nn.Conv1d(convolutions[1], convolutions[2], kernel_size, padding=padding, padding_mode='reflect'),
-            # nn.MaxPool1d(2, stride),
-            # nn.Conv1d(convolutions[2], convolutions[3], kernel_size, padding=padding, padding_mode='reflect'),
-            # nn.MaxPool1d(2, stride)
+            nn.Conv1d(num_channels, convolutions[0], kernel_size, padding=padding, padding_mode='reflect'),
+            nn.MaxPool1d(kernel_size, stride),
+            nn.Conv1d(convolutions[0], convolutions[1], kernel_size, padding=padding, padding_mode='reflect'),
+            nn.MaxPool1d(kernel_size, stride),
         )
 
         self.fully_connected = nn.Sequential(
             nn.Linear(self.image_size * convolutions[-1], hidden_layers[0]),
             nn.BatchNorm1d(num_features=num_channels),
-            nn.ReLU(),
+            nn.LeakyReLU(),
             nn.Linear(hidden_layers[0], output_size),
-            nn.BatchNorm1d(num_features=num_channels))
-
-        # TODO: fix the decoder: https://stackoverflow.com/questions/55033669/encoding-and-decoding-pictures-pytorch
-        # https://discuss.pytorch.org/t/visualize-feature-map/29597/6
-        # self.decoder = torch.nn.Sequential(
-            # nn.ConvTranspose1d(convolutions[3], convolutions[2], filter_size),
-            # nn.ReLU(),
-            # nn.ConvTranspose1d(convolutions[2], convolutions[1], filter_size),
-            # nn.ConvTranspose1d(convolutions[1], num_channels, filter_size))
+            nn.BatchNorm1d(num_features=num_channels),
+        )
 
     def forward(self, x):
         out = self.extractor(x)
-        out = out.view(-1, num_channels, self.image_size * convolutions[-1])  # -1 for number of aerofoils in batch
+        out = out.view(-1, num_channels, self.image_size * convolutions[-1])  # -1 for varying batch sizes
         out = self.fully_connected(out)
 
         return out  # you don't want to return a tuple, you must return a tensor
@@ -140,22 +131,25 @@ class ConvNet(nn.Module):
 # model, loss and optimiser
 model = ConvNet().to(device)
 criterion = nn.SmoothL1Loss()
-optimiser = torch.optim.SGD(model.parameters(), lr=learning_rate)
-scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimiser, 'min', patience=5, verbose=True)
+optimiser = torch.optim.Adam(model.parameters(), lr=learning_rate)
+scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimiser, 'min', patience=3, verbose=True)
 
-if find_LR:
+if find_LR and not torch.cuda.is_available():
     if learning_rate > 0.0001:
-        warnings.warn(f"Starting learning should be lowered from {learning_rate}")
+        warnings.warn(f"Selected initial learning rate too high.\nLearning rate changed to 0.0001")
+        optimiser = torch.optim.Adam(model.parameters(), lr=0.0001)
     lr_finder = LRFinder(model, optimiser, criterion, device=device)
-    lr_finder.range_test(train_loader, end_lr=100, num_iter=100)
+    lr_finder.range_test(train_loader, end_lr=500, num_iter=200)
     lr_finder.plot()  # to inspect the loss-learning rate graph
     lr_finder.reset()  # to reset the model and optimizer to their initial state
+    subprocess.Popen(["kill", "-9", f"{TB_process.pid}"])
     sys.exit("Learning rate plot finished")
 
 # structure of model
-for sample in train_loader:
-    writer.add_graph(model, sample[0].float())
-writer.close()
+if not torch.cuda.is_available():
+    for sample in train_loader:
+        writer.add_graph(model, sample[0].float())
+    writer.close()
 
 # training loop
 model.train()  # needed?
@@ -176,7 +170,7 @@ for epoch in range(num_epochs):
         train_loss.backward()
         optimiser.step()
 
-        if (epoch+1) % print_epoch == 0:  # at the 100th epoch:
+        if (epoch+1) % print_epoch == 0 or epoch == 0:  # at the 100th epoch:
             # calculate training loss
             running_train_loss += train_loss.item() * len(train_sample)   # loss.item() returns average loss per sample in batch
 
@@ -199,8 +193,9 @@ for epoch in range(num_epochs):
                 running_valid_loss /= len(valid_dataset) * 1
 
                 # print to TensorBoard
-                writer.add_scalar("training loss", running_train_loss, epoch)  # , epoch * len(train_dataset) + i
-                writer.add_scalar("validation loss", running_valid_loss, epoch)  # , epoch * len(train_dataset) + i
+                if not torch.cuda.is_available():
+                    writer.add_scalar("training loss", running_train_loss, epoch)  # , epoch * len(train_dataset) + i
+                    writer.add_scalar("validation loss", running_valid_loss, epoch)  # , epoch * len(train_dataset) + i
 
                 print(f"epoch {epoch+1}/{num_epochs}, batch {i+1}/{len(train_loader)}.\n"
                       f"Training loss = {running_train_loss:.4f}, "
@@ -209,7 +204,8 @@ for epoch in range(num_epochs):
                 scheduler.step(running_valid_loss)
                 running_train_loss = 0.
                 running_valid_loss = 0.
-writer.close()
+if not torch.cuda.is_available():
+    writer.close()
 torch.save(model.state_dict(), print_dir / "model.pkl")  # create pickle file
 
 # result = model.decode(model((valid_dataset[0])["coordinates"]).view(-1, num_channels, input_size))
@@ -220,8 +216,8 @@ model.eval()  # turn off batch normalisation and dropout
 with torch.no_grad():  # don't add gradients of test set to computational graph
     losses = {}
     running_test_loss = 0.
-    test_target_list = torch.tensor([])
-    test_predictions_list = torch.tensor([])
+    test_target_list = torch.tensor([]).to(device)
+    test_predictions_list = torch.tensor([]).to(device)
     for test_batch in test_loader:
         # data
         test_coords = test_batch[0].to(device)
@@ -277,9 +273,10 @@ if print_activations:
         plt.plot(act[idx])
 
 # kill TensorBoard
-TB_close = input("\nPress Y to kill TensorBoard...\n")
-if re.search('[yY]+', TB_close):
-    subprocess.Popen(["kill", "-9", f"{TB_process.pid}"])
-    print("TensorBoard closed.")
-else:
-    print("TensorBorad will remain open.")
+if not torch.cuda.is_available():
+    TB_close = input("\nPress Y to kill TensorBoard...\n")
+    if re.search('[yY]+', TB_close):
+        subprocess.Popen(["kill", "-9", f"{TB_process.pid}"])
+        print("TensorBoard closed.")
+    else:
+        print("TensorBorad will remain open.")
